+++ CHANGELOG.md
@@ -6,6 +6,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 #### Added
+#### Fixed
+#### Removed
+
+
+
+## [1.0.0] - 2017-03-10
+#### Added
 - **[AGENT]** Released Needle Agent
 - **[CORE]** iOS 10 Support
 - **[CORE]** Overhaul of the Core